+++ server/src/main/java/fcu/selab/progedu/setting/AssignmentSettings.java
@@ -1,0 +1,8 @@
+package fcu.selab.progedu.setting;
+
+import java.util.List;
+import java.io.IOException;
+
+public interface AssignmentSettings {
+  public void createAssignmentSetting(List<String> order,String name);
+}
