--- server/src/main/java/fcu/selab/progedu/setting/AssignmentTypeSettings.java
@@ -1,5 +1,0 @@
-package fcu.selab.progedu.setting;
-
-public interface AssignmentSettings {
-    public void createAssignmentSetting();
-}
